+++ pihotspot.sh
@@ -8,7 +8,7 @@ LOGNAME="pihotspot.log"
 # be sure to add a / at the end of the path
 LOGPATH="/var/log/"
 # Password for user root (MySql not system)
-MYSQL_PASSWORD="5a7sKncNxc"
+MYSQL_PASSWORD="pihotspot"
 # Name of the hotspot that will be visible for users/customers
 HOTSPOT_NAME="pihotspot"
 # IP of the hotspot
@@ -16,11 +16,9 @@ HOTSPOT_IP="192.168.10.1"
 # Network where the hotspot is located
 HOTSPOT_NETWORK="192.168.10.0"
 # Secret word for CoovaChilli
-COOVACHILLI_SECRETKEY="xBNRQLBUxe"
-# Password for radius database
-RADIUS_PASSWORD="8HXGwfUxYD"
+COOVACHILLI_SECRETKEY="change-me"
 # Secret word for FreeRadius
-FREERADIUS_SECRETKEY="sa8dtUkgp8"
+FREERADIUS_SECRETKEY="testing123"
 # WAN interface (the one with Internet)
 WAN_INTERFACE="eth0"
 # LAN interface (the one for the hotspot)
@@ -43,6 +41,15 @@ HASERL_URL="http://downloads.sourceforge.net/project/haserl/haserl-devel/haserl-
 # Haserl archive name based on the URL (keep the same version)
 HASERL_ARCHIVE="haserl-0.9.35"
 
+### PKG Vars ###
+PKG_MANAGER="apt-get"
+PKG_CACHE="/var/lib/apt/lists/"
+UPDATE_PKG_CACHE="${PKG_MANAGER} update"
+PKG_INSTALL="${PKG_MANAGER} --yes install"
+PKG_UPGRADE="${PKG_MANAGER} --yes upgrade"
+PKG_DIST_UPGRADE="apt dist-upgrade -y --force-yes"
+PKG_COUNT="${PKG_MANAGER} -s -o Debug::NoLocking=true upgrade | grep -c ^Inst || true"
+
 check_returned_code() {
     RETURNED_CODE=$@
     if [ $RETURNED_CODE -ne 0 ]; then
@@ -56,9 +63,9 @@ check_returned_code() {
 display_message() {
     MESSAGE=$@
     # Display on console
-    echo "** $MESSAGE"
+    echo "::: $MESSAGE"
     # Save it to log file
-    echo "** $MESSAGE" >> $LOGPATH$LOGNAME
+    echo "::: $MESSAGE" >> $LOGPATH$LOGNAME
 }
 
 execute_command() {
@@ -72,12 +79,12 @@ execute_command() {
 }
 
 prepare_logfile() {
-    echo "** Preparing log file"
+    echo "::: Preparing log file"
     if [ -f $LOGPATH$LOGNAME ]; then
-        echo "** Log file already exists. Creating a backup."
+        echo "::: Log file already exists. Creating a backup."
         execute_command "mv $LOGPATH$LOGNAME $LOGPATH$LOGNAME.`date +%Y%m%d.%H%M%S`"
     fi
-    echo "** Creating the log file"
+    echo "::: Creating the log file"
     execute_command "touch $LOGPATH$LOGNAME"
     display_message "Log file created : $LOGPATH$LOGNAME"
     display_message "Use command 'tail -f $LOGPATH$LOGNAME' in a new console to get installation details"
@@ -86,9 +93,28 @@ prepare_logfile() {
 prepare_install() {
     # Prepare the log file
     prepare_logfile
-    
+
     # Force IPv4 on APT resources
     execute_command "echo 'Acquire::ForceIPv4 \"true\";' | sudo tee /etc/apt/apt.conf.d/99force-ipv4" true "Updating APT config to force IPv4"
+}
+
+check_root() {
+    # Must be root to install the hotspot
+    echo ":::"
+    if [[ $EUID -eq 0 ]];then
+        echo "::: You are root - OK"
+    else
+        echo "::: sudo will be used for the install."
+        # Check if it is actually installed
+        # If it isn't, exit because the install cannot complete
+        if [[ $(dpkg-query -s sudo) ]];then
+            export SUDO="sudo"
+            export SUDOE="sudo -E"
+        else
+            echo "::: Please install sudo or run this as root."
+            exit 1
+        fi
+    fi
 }
 
 jumpto() {
@@ -98,29 +124,155 @@ jumpto() {
     exit
 }
 
+verifyFreeDiskSpace() {
+    # Needed free space
+    local required_free_kilobytes=500000
+    # If user installs unattended-upgrades we will check for 500MB free
+    echo ":::"
+    echo -n "::: Verifying free disk space ($required_free_kilobytes Kb)"
+    local existing_free_kilobytes=$(df -Pk | grep -m1 '\/$' | awk '{print $4}')
+
+    # - Unknown free disk space , not a integer
+    if ! [[ "${existing_free_kilobytes}" =~ ^([0-9])+$ ]]; then
+        echo ""
+        echo "::: Unknown free disk space!"
+        echo "::: We were unable to determine available free disk space on this system."
+        echo "::: You may continue with the installation, however, it is not recommended."
+        read -r -p "::: If you are sure you want to continue, type YES and press enter :: " response
+        case $response in
+            [Y][E][S])
+                ;;
+            *)
+                echo "::: Confirmation not received, exiting..."
+                exit 1
+                ;;
+        esac
+    # - Insufficient free disk space
+    elif [[ ${existing_free_kilobytes} -lt ${required_free_kilobytes} ]]; then
+        echo ""
+        echo "::: Insufficient Disk Space!"
+        echo "::: Your system appears to be low on disk space. Pi-HotSpot recommends a minimum of $required_free_kilobytes KiloBytes."
+        echo "::: You only have ${existing_free_kilobytes} KiloBytes free."
+        echo ":::"
+        echo "::: If this is a new install on a Raspberry Pi you may need to expand your disk."
+        echo "::: Try running 'sudo raspi-config', and choose the 'expand file system option'"
+        echo ":::"
+        echo "::: After rebooting, run this installation again. (curl -L https://raw.githubusercontent.com/pihomeserver/Pi-Hotspot-Script/master/pihotspot.sh | bash)"
+
+        echo "Insufficient free space, exiting..."
+        exit 1
+    else
+        echo " - OK"
+    fi
+}
+
+update_package_cache() {
+  timestamp=$(stat -c %Y ${PKG_CACHE})
+  timestampAsDate=$(date -d @"${timestamp}" "+%b %e")
+  today=$(date "+%b %e")
+
+  if [ ! "${today}" == "${timestampAsDate}" ]; then
+    #update package lists
+    echo ":::"
+    #echo -n "::: ${PKG_MANAGER} update has not been run today. Running now..."
+    if command -v debconf-apt-progress &> /dev/null; then
+        $SUDO debconf-apt-progress -- ${UPDATE_PKG_CACHE}
+    else
+        $SUDO ${UPDATE_PKG_CACHE} &> /dev/null
+    fi
+    #echo " done!"
+  fi
+}
+
+notify_package_updates_available() {
+  echo ":::"
+  echo -n "::: Checking ${PKG_MANAGER} for upgraded packages...."
+  updatesToInstall=$(eval "${PKG_COUNT}")
+  echo " done!"
+  echo ":::"
+  if [[ ${updatesToInstall} -eq "0" ]]; then
+    echo "::: Your system is up to date! Continuing with Pi-Hotspot installation..."
+  else
+    echo "::: There are ${updatesToInstall} updates available for your system!"
+    echo ":::"
+    execute_command "apt-get upgrade -y --force-yes" true "Upgrading the packages. Please be patient."
+  fi
+}
+
+package_check_install() {
+    dpkg-query -W -f='${Status}' "${1}" 2>/dev/null | grep -c "ok installed" || ${PKG_INSTALL} "${1}"
+}
+
+PIHOTSPOT_DEPS_START=( apt-transport-https )
+PIHOTSPOT_DEPS_WIFI=( apt-utils firmware-brcm80211 firmware-ralink firmware-realtek )
+PIHOTSPOT_DEPS=( wget build-essential grep whiptail debconf-utils git libjson-c-dev haserl gengetopt devscripts libtool bash-completion autoconf automake hostapd php5-mysql php-pear php5-gd php-db php5-fpm libgd2-xpm-dev libpcrecpp0 libxpm4 nginx php5-xcache debhelper libssl-dev libcurl4-gnutls-dev mysql-server freeradius freeradius-mysql gcc make libnl1 libnl-dev pkg-config iptables git libjson-c-dev haserl gengetopt devscripts libtool bash-completion autoconf automake )
+
+install_dependent_packages() {
+  declare -a argArray1=("${!1}")
+
+  if command -v debconf-apt-progress &> /dev/null; then
+    $SUDO debconf-apt-progress -- ${PKG_INSTALL} "${argArray1[@]}"
+  else
+    for i in "${argArray1[@]}"; do
+      echo -n ":::    Checking for $i..."
+      $SUDO package_check_install "${i}" &> /dev/null
+      echo " installed!"
+    done
+  fi
+}
+
+check_root
+
+verifyFreeDiskSpace
+
 prepare_install
 
-# Use the following instructions to help debug the script and bypass some instructions that have been already executed
-# 1) Uncomment both lines (jumpto and nextstep)
-# 2) Move the line "nextstep:" right before the first instruction that you want to be executed
+update_package_cache
 
-#jumpto "nextstep"
-#nextstep:
+notify_package_updates_available
 
-execute_command "apt-get update" true "Updating system"
-execute_command "apt-get upgrade -y" true "Upgrading all packages"
-execute_command "apt-get install -y --force-yes apt-transport-https" true "Adding HTTPS support for apt-get (Raspbian Lite compatiblity)"
+install_dependent_packages PIHOTSPOT_DEPS_START[@]
 
-execute_command "ifconfig -a | grep $LAN_INTERFACE" false "Checking if wlan0 interface already exists"
+execute_command "/sbin/lsmod | grep tun" false "Checking for tun module"
+if [ $COMMAND_RESULT -ne 0 ]; then
+    display_message "Insert tun module if existing (for Raspbian Jessie Lite)"
+    find /lib/modules/ -iname "tun.ko.gz" -exec /sbin/insmod {} \;
+    check_returned_code $?
+
+    display_message "Modprobe module (no check - useless if already loaded)"
+    /sbin/modprobe tun
+
+    execute_command "/sbin/lsmod | grep tun" false "Checking for tun module"
+    if [ $COMMAND_RESULT -ne 0 ]; then
+        display_message "Unable to get tun module up. Please solve before running the script again."
+        display_message "If your distribution has been upgraded you should try to reboot first."
+        exit 1
+    fi
+fi
+
+execute_command "/sbin/ifconfig -a | grep $LAN_INTERFACE" false "Checking if wlan0 interface already exists"
 if [ $COMMAND_RESULT -ne 0 ]; then
     display_message "Wifi interface not found. Upgrading the system first"
+
     execute_command "apt dist-upgrade -y --force-yes" true "Upgrading the distro. Be patient"
-    execute_command "apt-get install apt-utils firmware-brcm80211 -y --force-yes" true "Install Wifi firmware"
+
+    install_dependent_packages PIHOTSPOT_DEPS_WIFI[@]
+
     display_message "Please reboot and run the script again"
     exit 1
 fi
 
-display_message "Update interface configuration"
+execute_command "echo 'mysql-server mysql-server/root_password password $MYSQL_PASSWORD' | debconf-set-selections" true "Adding MySql password"
+execute_command "echo 'mysql-server mysql-server/root_password_again password $MYSQL_PASSWORD' | debconf-set-selections" true "Adding MySql password (confirmation)"
+
+install_dependent_packages PIHOTSPOT_DEPS[@]
+
+notify_package_updates_available
+
+execute_command "service mysql restart" true "Starting MySql service"
+
+execute_command "grep $LAN_INTERFACE /etc/network/interfaces" false "Update interface configuration"
+if [ $COMMAND_RESULT -ne 0 ]; then
 cat >> /etc/network/interfaces << EOT
 
 auto $LAN_INTERFACE
@@ -131,29 +283,19 @@ iface $LAN_INTERFACE inet static
     network $HOTSPOT_NETWORK
     post-up echo 1 > /proc/sys/net/ipv4/ip_forward
 EOT
-check_returned_code $?
+    check_returned_code $?
+fi
 
 execute_command "ifup $WAN_INTERFACE" true "Activating the WAN interface"
 execute_command "ifup $LAN_INTERFACE" true "Activating the LAN interface"
 
-execute_command "apt-get install -y --force-yes debconf-utils" true "Installing debconf tools"
-
-execute_command "echo 'mysql-server mysql-server/root_password password $MYSQL_PASSWORD' | debconf-set-selections" true "Adding MySql password"
-#execute_command "debconf-set-selections <<< 'mysql-server mysql-server/root_password password $MYSQL_PASSWORD'" true "Adding MySql password"
-execute_command "echo 'mysql-server mysql-server/root_password_again password $MYSQL_PASSWORD' | debconf-set-selections" true "Adding MySql password (confirmation)"
-#execute_command "debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password $MYSQL_PASSWORD'" true "Adding MySql password (confirmation)"
-execute_command "apt-get install -y --force-yes debhelper libssl-dev libcurl4-gnutls-dev mysql-server freeradius freeradius-mysql gcc make libnl1 libnl-dev pkg-config iptables" true "Installing MySql, freeradius, iptables and other dependencies"
-
 display_message "Creating freeradius database"
+echo 'drop database if exists radius;' | mysql -u root -p$MYSQL_PASSWORD
 echo 'create database radius;' | mysql -u root -p$MYSQL_PASSWORD
 check_returned_code $?
 
 display_message "Installing freeradius schema"
 mysql -u root -p$MYSQL_PASSWORD radius < /etc/freeradius/sql/mysql/schema.sql
-check_returned_code $?
-
-display_message "Update radius database password in user creation"
-sed -i "s/radpass/$RADIUS_PASSWORD/g" /etc/freeradius/sql/mysql/admin.sql
 check_returned_code $?
 
 display_message "Creating administrator privileges"
@@ -178,23 +320,11 @@ check_returned_code $?
 
 execute_command "service freeradius stop" true "Stoping freeradius service to update the configuration"
 
-display_message "Activating SQL authentication"
-sed -i '/^#.*sql$/s/^#//g' /etc/freeradius/sites-available/default
-check_returned_code $?
-
-#display_message "Activating CoA site"
-#ln -s /etc/freeradius/sites-available/coa /etc/freeradius/sites-enabled/coa
-#check_returned_code $?
-
-execute_command "freeradius -C" true "Checking freeradius configuration"
-execute_command "service freeradius start" true "Starting freeradius service"
-
 display_message "Activating IP forwarding"
 sed -i '/^#net\.ipv4\.ip_forward=1$/s/^#//g' /etc/sysctl.conf
 check_returned_code $?
 execute_command "/etc/init.d/networking restart" true "Restarting network service to take IP forwarding into account"
 
-execute_command "apt-get install -y --force-yes git libjson-c-dev haserl gengetopt devscripts libtool bash-completion autoconf automake" true "Installing compilation tools"
 execute_command "cd /usr/src && git clone $COOVACHILLI_ARCHIVE" true "Cloning CoovaChilli project"
 
 execute_command "cd /usr/src/coova-chilli && dpkg-buildpackage -us -uc" true "Building CoovaChilli package"
@@ -260,8 +390,6 @@ if [ $COMMAND_RESULT -ne 0 ]; then
     exit 1
 fi
 
-execute_command "apt-get install -y --force-yes hostapd" true "Installing hostapd"
-
 display_message "Creating configuration file for hostapd"
 echo 'DAEMON_CONF="/etc/hostapd/hostapd.conf"' >> /etc/default/hostapd
 check_returned_code $?
@@ -281,30 +409,11 @@ check_returned_code $?
 
 execute_command "service hostapd start" true "Starting hostapd service"
 
-execute_command "apt-get install -y --force-yes php5-mysql php-pear php5-gd php-db php5-fpm libgd2-xpm-dev libpcrecpp0 libxpm4 nginx php5-xcache" true "Installing Nginx webserver with PHP support"
-
 execute_command "cd /usr/src && wget $DALORADIUS_ARCHIVE" true "Downloading daloradius"
 
 execute_command "cd /usr/src && tar zxvf download -C /usr/share/nginx/html/" true "Uncompressing daloradius archive"
 
 execute_command "mv /usr/share/nginx/html/daloradius-0.9-9 /usr/share/nginx/html/daloradius" true "Renaming daloradius folder"
-
-execute_command "service freeradius stop" true "Stopping freeradius service"
-
-execute_command "mv /etc/freeradius/sql/mysql/counter.conf /etc/freeradius/sql/mysql/counter.conf.bak && cp /usr/share/nginx/html/daloradius/contrib/configs/freeradius-2.1.8/cfg1/raddb/sql/mysql/counter.conf /etc/freeradius/sql/mysql/counter.conf" true "Updating /etc/freeradius/sql/mysql/counter.conf"
-#below is needed because counter.conf contains duplicate entry "sqlcounter noresetcounter", this will remove the first one.
-sed -i "110,117d" /etc/freeradius/sql/mysql/counter.conf
-check_returned_code $?
-
-execute_command "mv /etc/freeradius/sites-available/default /etc/freeradius/sites-available/default.bak && cp /usr/share/nginx/html/daloradius/contrib/configs/freeradius-2.1.8/cfg1/raddb/sites-available/default /etc/freeradius/sites-available/default" true "Updating /etc/freeradius/sites-available/default"
-
-execute_command "mv /etc/freeradius/sql.conf /etc/freeradius/sql.conf.bak && cp /usr/share/nginx/html/daloradius/contrib/configs/freeradius-2.1.8/cfg1/raddb/modules/sql.conf /etc/freeradius/sql.conf" true "Updating /etc/freeradius/sql.conf"
-
-display_message "Update radius database password in /etc/freeradius/sql.conf"
-sed -i "s/radpass/$RADIUS_PASSWORD/g" /etc/freeradius/sql.conf
-check_returned_code $?
-
-execute_command "service freeradius start" true "Starting freeradius service"
 
 display_message "Loading daloradius configuration into MySql"
 mysql -u root -p$MYSQL_PASSWORD radius < /usr/share/nginx/html/daloradius/contrib/db/fr2-mysql-daloradius-and-freeradius.sql
@@ -324,7 +433,7 @@ display_message "Configuring daloradius DB user name"
 sed -i "s/\$configValues\['CONFIG_DB_USER'\] = 'root';/\$configValues\['CONFIG_DB_USER'\] = 'radius';/g" /usr/share/nginx/html/daloradius/library/daloradius.conf.php
 check_returned_code $?
 display_message "Configuring daloradius DB user password"
-sed -i "s/\$configValues\['CONFIG_DB_PASS'\] = '';/\$configValues\['CONFIG_DB_PASS'\] = '$RADIUS_PASSWORD';/g" /usr/share/nginx/html/daloradius/library/daloradius.conf.php
+sed -i "s/\$configValues\['CONFIG_DB_PASS'\] = '';/\$configValues\['CONFIG_DB_PASS'\] = 'radpass';/g" /usr/share/nginx/html/daloradius/library/daloradius.conf.php
 check_returned_code $?
 
 display_message "Building NGINX configuration (default listen port : 80)"
@@ -352,6 +461,19 @@ check_returned_code $?
 
 execute_command "nginx -t" true "Checking Nginx configuration file"
 
+execute_command "mv /etc/freeradius/sql/mysql/counter.conf /etc/freeradius/sql/mysql/counter.conf.bak && cp /usr/share/nginx/html/daloradius/contrib/configs/freeradius-2.1.8/cfg1/raddb/sql/mysql/counter.conf /etc/freeradius/sql/mysql/counter.conf" true "Updating /etc/freeradius/sql/mysql/counter.conf"
+
+display_message "Remove duplicate entry 'sqlcounter noresetcounter'"
+sed -i "110,117d" /etc/freeradius/sql/mysql/counter.conf
+check_returned_code $?
+
+execute_command "mv /etc/freeradius/sites-available/default /etc/freeradius/sites-available/default.bak && cp /usr/share/nginx/html/daloradius/contrib/configs/freeradius-2.1.8/cfg1/raddb/sites-available/default /etc/freeradius/sites-available/default" true "Updating /etc/freeradius/sites-available/default"
+
+execute_command "mv /etc/freeradius/sql.conf /etc/freeradius/sql.conf.bak && cp /usr/share/nginx/html/daloradius/contrib/configs/freeradius-2.1.8/cfg1/raddb/modules/sql.conf /etc/freeradius/sql.conf" true "Updating /etc/freeradius/modules/sql.conf"
+
+execute_command "freeradius -C" true "Checking freeradius configuration"
+execute_command "service freeradius start" true "Starting freeradius service"
+
 execute_command "service nginx restart" true "Restarting Nginx"
 
 execute_command "service hostapd restart" true "Restarting hostapd"
@@ -374,3 +496,4 @@ display_message "  (login : administrator / password : radius)"
 #echo "exit 0" >> /etc/rc.local
 
 exit 0
+